sdaSD

+++ GestionRutas/IngresoRutas.java
@@ -140,19 +140,32 @@ public class IngresoRutas {
         if(r.controlNum(id_Ruta)){
             deleteRuta(Integer.parseInt(id_Ruta));
         }else if(!r.controlNum(id_Ruta)){
-            System.out.println("El valor ingresado no concuerda con los id's\nPorfavor vuelvalo a ingresar");
+            System.out.println("El valor ingresado debe ser númerico\nPorfavor vuelvalo a ingresar");
             MenudeleteRuta();
         }
     }
 
     public boolean deleteRuta(int id_Rutas) {
         try {
-            conexion.setP(c.prepareStatement("DELETE FROM Rutas WHERE Id_Rutas = '" + id_Rutas + "'"));
-            conexion.getP().executeUpdate();
+            conexion.setP(c.prepareStatement("SELECT *FROM Rutas WHERE Id_Rutas" + id_Rutas + "'"));
+                conexion.setRs(conexion.getP().executeQuery());
+                if (conexion.getRs().next()) {
+                    try {
+                        conexion.setP(c.prepareStatement("DELETE FROM Rutas WHERE Id_Rutas = '" + id_Rutas + "'"));
+                        conexion.getP().executeUpdate();
+                    } catch (SQLException e) {
+                        System.out.println(" === ERROR DE INGRESO EN BD ===");
+                        System.out.println(e.getMessage());
+                    }
+                }else{
+                    System.out.println("El valor ingresado no concuerda con los id's\nPorfavor vuelvalo a ingresar");
+                    MenudeleteRuta();
+                }
+
         } catch (SQLException e) {
             System.out.println(" === ERROR DE INGRESO EN BD ===");
-            System.out.println(e.getMessage());
         }
+        
 
         return true;
     }
